requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
bleach==6.2.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-filter==24.3
django-otp==1.5.4
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
Markdown==3.7
msgpack==1.1.0
pillow==11.0.0
prompt_toolkit==3.0.48
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyotp==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zope.interface==7.2
